Save new data source + credentials upon Connect.
Use python ctypes module to invoke SQLConfigDataSource when connecting
to a new data source (the data source can exist and the setup dll in the
driver will update or ask to update it). This may invoke the GUI found
in the driver setup dll.

Use python keyring module to save credentials if present for a new data
source name.