Add more signals to widgets, add Manage button
Add more signals to widget to myke the UI interactive and populate
driver name / dsn name automatically when the list is clicked.

Retrieve stored credentials for a DSN, when clicked from the list,
and clear them when clicking other Driver or DSN. Check for user input
and do not override credentials typed by user.

Add button to run ODBC Data Source Administrator installed in the
system.