Add default autocommit mode, add 1000 rows limit.
Restore the New Tab... button ("+" button) at the end of the SQL editor
tab bar.

Add special case for 'DBASE' driver to skip call to SQLProcedures
(unsupported) - should check use of exceptions instead if available.

If result rows exceed 1000, show limit of 1000 in the Result tab label.

Ensure empty SQL editor conent is still saved to a data file when
closing the DatabaseView, to prevent "File not found" error on next
open of the same connection.