Separate enumeration of user and system DSNs
Use "odbc:" prefix instead of "odbc-client:" for saving credentials.

Move calls to ODBCInst.Init() right in the module file, so the module is
always initiailzed when imported. Same for calls to ODBC.Init() for the
new ODBC module.

Add new ODBC.py module with a couple of ctypes binding for SQL/CLI C
functions:
    - SQLAllocEnv
    - SQLFreeEnv
    - SQLSetEnvAttr
    - SQLDataSources

Use new functions in ODBC module to separate enumeration of user DSNs
and system DSNs.